requirements.txt
gunicorn==19.6.0
python-telegram-bot
